+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 update/optimize
 
+- update ResourceOp, update TypeHint
+
 add
 
 fix